configured base redux
just store with sample test reducer
hot reload config
reogranised the components
redux logger and dev tools
