+++ client/src/containers/Root/Root.js
@@ -1,0 +1,5 @@
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./Root.prod')
+} else {
+  module.exports = require('./Root.dev')
+}
