+++ client/src/containers/Root/Root.prod.js
@@ -1,0 +1,19 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { Provider } from 'react-redux'
+import { Route } from 'react-router-dom'
+import App from '../App/App'
+
+const Root = ({ store }) => (
+  <Provider store={store}>
+    <div>
+      <App />
+    </div>
+  </Provider>
+)
+
+Root.propTypes = {
+  store: PropTypes.object.isRequired,
+}
+
+export default Root
