+++ client/src/middleware/api.js
@@ -1,0 +1,11 @@
+
+export const Schemas = {}
+
+// Action key that carries API call info interpreted by this Redux middleware.
+export const CALL_API = 'Call API'
+
+// A Redux middleware that interprets actions with CALL_API info specified.
+// Performs the call and promises when such actions are dispatched.
+export default store => next => action => {
+  return new Promise.resolve();
+}
