+++ client/src/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+
+
+let reducer = combineReducers({
+  test: () => {
+    return '';
+  }
+});
+
+export default reducer;
