+++ client/src/store/configureStore.prod.js
@@ -1,0 +1,12 @@
+import { createStore, applyMiddleware } from 'redux'
+import thunk from 'redux-thunk'
+import api from '../middleware/api'
+import rootReducer from '../reducers'
+
+const configureStore = preloadedState => createStore(
+  rootReducer,
+  preloadedState,
+  applyMiddleware(thunk, api)
+)
+
+export default configureStore
